# docs/openapi.yaml
openapi: 3.0.3
info:
  description: |
    # Introduction

    Welcome to the Branda API documentation! The Branda API is used by both the Branda app and the Branda Admin
    Panel. While the API has some publicly-accessible routes, most require an API key to successfully complete an 
    operation (see the [Authentication and Security](#section/Authentication-and-Security) section for more details).

    # Authentication and Security

    Petstore offers two forms of authentication:
      - API Key
      - DB Key

    The API Key is a general-use token specified in the header of almost all requests sent to our API. Because the majority
    of this API deals with sensitive information related to service tracking, push notifications, or backend operations, API
    security is a crucial part of our development. Some routes are publicly-available, and will not have the API key marked as 
    *required* in the header parameters. If you would like more information regarding this key, and possibly gaining access
    for personal development or integration with other applications, please reach out to the Branda team.

    The DB Key is specific to `Admin`-related routes, and will be marked as such in routes that use it. This key will
    not be given out under any circumstances.

     # OpenAPI Specification
    This API is documented in **OpenAPI format**.
    It was **extended** to illustrate features of [generator-openapi-repo](https://github.com/Rebilly/generator-openapi-repo)
...
paths:
...
  /api/schedules/{week}/{id}:
    parameters:
      - in: header
        name: api_token
        description: API Key for sender verification and general security
        required: true
        schema:
          type: string
      - in: path
        name: week
        description: Week of the year to get schedules for (0-based)
        required: true
        schema:
          type: integer
          format: int32
        example: 30
      - in: path
        name: id
        description: Employee ID to get the schedule for
        required: true
        schema:
          type: string
        example: 5b9b46afd5e64959c00eb455
    get:
      tags:
        - Schedules
      summary: Get Week Schedule for A Specific Place
      description: Fetches the week data for a specific schedule based on week and place id
      responses:
        200:
          description: Successfully fetched specific place's schedule on a specific week
          content:
            application/json:
              schema:
                type: object
                properties:
                  week:
                    description: Week selected
                    type: integer
                    format: int32
                    example: 30
                  sunday:
                    description: Sunday schedule for the specific week
                    type: string
                    example: "12:00pm-7:00pm"
                  monday:
                    description: Monday schedule for the specific week
                    type: string
                    example: "11:00am-11:00pm"
                  tuesday:
                    description: Tuesday schedule for the specific week
                    type: string
                    example: "11:00am-11:00pm"
                  wednesday:
                    description: Wednesday schedule for the specific week
                    type: string
                    example: "11:00am-11:00pm"
                  thursday:
                    description: Thursday schedule for the specific week
                    type: string
                    example: "11:00am-11:00pm"
                  friday:
                    description: Friday schedule for the specific week
                    type: string
                    example: "11:00am-8:00pm"
                  saturday:
                    description: Saturday schedule for the specific week
                    type: string
                    example: "12:00am-8:00pm"
        404:
          description: Place schedule not found, indicative of non-existent employee ID
        405:
          description: Incorrect HTTP method
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/mongoError"
    patch:
      tags:
        - Schedules
      summary: Update Week Schedule for A Specific Place
      description: Updates the week data for a specific schedule based on week and place id
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - sunday
                - monday
                - tuesday
                - wednesday
                - thursday
                - friday
                - saturday
              properties:
                sunday:
                  description: New Sunday schedule for the place on a specific week
                  type: string
                  format: "h:mm[aApP][mM]-h:mm[aApP][mM]"
                  example: "12:00pm-5:00pm"
                monday:
                  description: New Monday schedule for the place on a specific week
                  type: string
                  format: "h:mm[aApP][mM]-h:mm[aApP][mM]"
                  example: "9:00am-5:00pm"
                tuesday:
                  description: New Tuesday schedule for the place on a specific week
                  type: string
                  format: "h:mm[aApP][mM]-h:mm[aApP][mM]"
                  example: "9:00am-5:00pm"
                wednesday:
                  description: New Wednesday schedule for the place on a specific week
                  type: string
                  format: "h:mm[aApP][mM]-h:mm[aApP][mM]"
                  example: "9:00am-5:00pm"
                thursday:
                  description: New Thursday schedule for the place on a specific week
                  type: string
                  format: "h:mm[aApP][mM]-h:mm[aApP][mM]"
                  example: "9:00am-5:00pm"
                friday:
                  description: New Friday schedule for the place on a specific week
                  type: string
                  format: "h:mm[aApP][mM]-h:mm[aApP][mM]"
                  example: "9:00am-5:00pm"
                saturday:
                  description: New Saturday schedule for the place on a specific week
                  type: string
                  format: "h:mm[aApP][mM]-h:mm[aApP][mM]"
                  example: "11:00am-5:00pm"
      responses:
        200:
          description: Successfully patched week schedule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceSchedule'
        405:
          description: Incorrect HTTP method
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/mongoError"
...
components:
  responses:
    mongoError:
      type: object
      properties:
        err:
          description: Error thrown
          type: string
          example: |
            MongooseError: The `uri` parameter to `openUri()` must be a string, got "undefined". 
            Make sure the first parameter to `mongoose.connect()` or `mongoose.createConnection()` is a string.
  schemas:
    PlaceSchedule:
      type: object
      properties:
          _id:
            description: Mongoose-assigned Object Id
            type: ObjectId
          emp_id:
            type: string
            description: Employee ID for the new place schedule
          Name:
            type: string
            description: Name for the new place schedule
          group:
            type: string
            description: Group for the new place schedule
          weeks:
            type: array
            description: Array with the schedules for each week of the year.
            items:
              type: object
              properties:
                sunday:
                  description: Schedule on a Sunday of a given week (based on the array index)
                  type: string
                  format: "h:mm[aApP][mM]-h:mm[aApP][mM]"
                monday:
                  description: Schedule on a Monday of a given week (based on the array index)
                  type: string
                  format: "h:mm[aApP][mM]-h:mm[aApP][mM]"
                tuesday:
                  description: Schedule on a Tuesday of a given week (based on the array index)
                  type: string
                  format: "h:mm[aApP][mM]-h:mm[aApP][mM]"
                wednesday:
                  description: Schedule on a Wednesday of a given week (based on the array index)
                  type: string
                  format: "h:mm[aApP][mM]-h:mm[aApP][mM]"
                thursday:
                  description: Schedule on a Thursday of a given week (based on the array index)
                  type: string
                  format: "h:mm[aApP][mM]-h:mm[aApP][mM]"
                friday:
                  description: Schedule on a Friday of a given week (based on the array index)
                  type: string
                  format: "h:mm[aApP][mM]-h:mm[aApP][mM]"
                saturday:
                  description: Schedule on a Saturday of a given week (based on the array index)
                  type: string
                  format: "h:mm[aApP][mM]-h:mm[aApP][mM]"
          sunday:
            description: General Sunday schedule
            type: string
          monday:
            description: General Sunday schedule
            type: string
          tuesday:
            description: General Sunday schedule
            type: string
          wednesday:
            description: General Sunday schedule
            type: string
          thursday:
            description: General Sunday schedule
            type: string
          friday:
            description: General Sunday schedule
            type: string
          saturday:
            description: General Sunday schedule
            type: string
          active:
            description: 1 represents an active place schedule, 0 represents an inactive place schedule
            type: integer
            format: int32
          __v:
            description: Mongoose-tracked internal revision of the document
            type: integer
            format: int32
      example:
          _id: 6bcb285dabd114f592f423a6c
          emp_id: 5b9b485dabd11f5920423a1c
          Name: The Stein
          group: Dining
          weeks: []
          sunday: '12:00pm-5:00pm'
          monday: '9:00am-5:00pm'
          tuesday: '9:00am-5:00pm'
          wednesday: '9:00am-5:00pm'
          thursday: '9:00am-5:00pm'
          friday: '9:00am-5:00pm'
          saturday: '11:00am-5:00pm'
          active: 1
          __v: 0